+++ csatar-plugins/csatar/Plugin.php
@@ -8,6 +8,7 @@ use Csatar\Csatar\Models\MandateType;
 use Csatar\Csatar\Models\Scout;
 use Csatar\Csatar\Classes\ContentPageSearchProvider;
 use Csatar\Csatar\Classes\OrganizationSearchProvider;
+use Db;
 use Event;
 use Input;
 use Media\Classes\MediaLibrary;
@@ -303,5 +304,15 @@ class Plugin extends PluginBase
         ];
     }
 
-
+    public function registerSchedule($schedule)
+    {
+        $schedule->call(function () {
+            Db::select(
+                "UPDATE csatar_csatar_scouts
+                SET family_name = '" . Scout::NAME_DELETED_INACTIVITY . "', given_name = ''
+                WHERE inactivated_at < DATE_SUB(NOW(), INTERVAL 5 YEAR) AND family_name <> '" . Scout::NAME_DELETED_INACTIVITY . "';"
+            );
+        })
+            ->daily();
+    }
 }