+++ csatar-plugins/csatar/models/Troop.php
@@ -157,8 +157,8 @@ class Troop extends OrganizationBase
                 $patrol->ignoreValidation = true;
                 $patrol->forceSave();
             }
-            foreach (Scout::where(['troop_id' => $this->id, 'is_active' => Status::ACTIVE])->get() as $scout) {
-                $scout->is_active = Status::INACTIVE;
+            foreach (Scout::where('troop_id', $this->id)->whereNull('inactivated_at')->get() as $scout) {
+                $scout->inactivated_at = date('Y-m-d H:i:s');
                 $scout->ignoreValidation = true;
                 $scout->forceSave();
             }