+++ csatar-plugins/knowledgerepository/updates/SeederData.php
@@ -2,6 +2,7 @@
 
 use Csatar\KnowledgeRepository\Models\AccidentRiskLevel;
 use Csatar\KnowledgeRepository\Models\GameDevelopmentGoal;
+use Csatar\KnowledgeRepository\Models\MethodologyType;
 use Csatar\KnowledgeRepository\Models\Tool;
 use Csatar\KnowledgeRepository\Models\Headcount;
 use Csatar\KnowledgeRepository\Models\Duration;
@@ -272,6 +273,10 @@ class SeederData extends Seeder
                 'model' => 'Csatar\KnowledgeRepository\Models\Game',
             ],
             [
+                'title' => 'Módszertan',
+                'model' => 'Csatar\KnowledgeRepository\Models\Methodology',
+            ],
+            [
                 'title' => 'Próbarendszer',
                 'model' => 'Csatar\KnowledgeRepository\Models\TrialSystem',
             ],
@@ -415,6 +420,20 @@ class SeederData extends Seeder
             'Ember és környezete',
             'Isten és ember',
         ],
+        'methodologyTypes' => [
+            [
+                'name' => 'Új (eddig az őrs még sosem találkozott az átadott anyaggal)',
+                'sort_order' => 1
+            ],
+            [
+                'name' => 'Régi/ismétlés (ismétlés, az őrs már találkozott a megnevezett anyaggal)',
+                'sort_order' => 2
+            ],
+            [
+                'name' => 'Mindkettő',
+                'sort_order' => 3
+            ],
+        ]
     ];
 
     public function run()
@@ -490,6 +509,15 @@ class SeederData extends Seeder
         foreach ($this::DATA['forms'] as $formData) {
             $form = Form::firstOrNew($formData);
             $form->save();
+        }
+        // Methodology Types
+
+        foreach ($this::DATA['methodologyTypes'] as $methodologyTypeData) {
+            $methodologyType = MethodologyType::firstOrNew([
+                'name' => $methodologyTypeData['name'],
+            ]);
+            $methodologyType->sort_order = $methodologyTypeData['sort_order'];
+            $methodologyType->save();
         }
 
         // Trial System Topic